Fixes #984 (#985)
