+++ android/app/src/androidTest/java/org/fossasia/openevent/DatabaseTest.java
@@ -59,7 +59,7 @@ public class DatabaseTest extends AndroidTestCase {
         Timber.tag("Micro").d(microlocationQuery);
         queries.add(microlocationQuery);
         Session session = new Session(5, "abcd", "abc", "abcdefgh", "sdfjs dsjfnjs",
-                "2015-06-05T00:00:00", "2015-06-06T00:00:00", "2015-06-06", "1", new org.fossasia.openevent.data.parsingextras.Track(6, "kids"), "0", new org.fossasia.openevent.data.parsingextras.Microlocation(4, "moscone centre"));
+                "2015-06-05T00:00:00", "2015-06-06T00:00:00", "2015-06-06", "1", new org.fossasia.openevent.data.parsingExtra.Track(6, "kids"), "0", new org.fossasia.openevent.data.parsingExtra.Microlocation(4, "moscone centre"));
 
         String sessionQuery = session.generateSql();
         Timber.tag("Session").d(sessionQuery);
